+++ jsk_2016_01_baxter_apc/scripts/interface_generator_stow.py
@@ -47,7 +47,6 @@ import copy
 import json
 import random
 import os
-import argparse
 import jsk_apc2016_common
 
 # define our bin and item names to use
@@ -193,23 +192,26 @@ class InterfaceGeneratorStow():
         return item_name
 
     # write the dictionary to the appropriately names json file
-    def write_json(self,version):
+    def write_json(self):
         data = {'bin_contents': self.bin_contents, 'work_order': self.tote_contents}
         this_dir = os.path.dirname(os.path.abspath(__file__))
         output_dir = os.path.join(this_dir,'../json')
         if not os.path.exists(output_dir):
             os.mkdir(output_dir)
         os.chdir(output_dir)
-        with open('stow_layout_'+version+'.json', 'w') as outfile:
-            json.dump(data, outfile, sort_keys=True, indent=4, separators=(',',': '))
-        print('stow_layout_'+version+'.json generated at ../json')
-        os.chdir(this_dir)
+        i=1
+        while True:
+            filename  = 'stow_layout_'+str(i)+'.json' 
+            if os.path.isfile(filename):
+                i +=1
+            else:
+                with open(filename, 'w') as outfile:
+                    json.dump(data, outfile, sort_keys=True, indent=4, separators=(',',': '))
+                print(filename+ ' generated at ../json')
+                os.chdir(this_dir)
+                break
 
 if __name__ == "__main__":
-    parser = argparse.ArgumentParser()
-    parser.add_argument("version")
-    args = parser.parse_args()
-    version = args.version
     interface_stow = InterfaceGeneratorStow()
     interface_stow.run()
-    interface_stow.write_json(version)
+    interface_stow.write_json()
